requirements.txt
Django==2.2.5
django-debug-toolbar==2.0
django-filter==2.2.0
djangorestframework==3.10.3
mysqlclient==1.4.4
Pillow==6.1.0
python-dotenv==0.10.3
pytz==2019.2
sqlparse==0.3.0
